requirements.txt

numpy
opencv-python
pytesseract
spacy>=3.0.0
python-pptx
